refactoring to rename AbstractSelectOptionFeederComponent to ASelectOptionFeeder
