+++ packages/ng/src/app/select/option/feeder/index.ts
@@ -1,3 +1,2 @@
 export {LuOptionFeederDirective} from './select-option-feeder.directive';
-export {AbstractSelectOptionFeederComponent} from './select-option-feeder.component';
-export {ISelectOptionFeeder} from './select-option-feeder.model';
+export {ASelectOptionFeeder, ISelectOptionFeeder} from './select-option-feeder.model';
