+++ packages/ng/src/app/select/option/feeder/select-option-feeder.model.ts
@@ -31,3 +31,46 @@ export interface ISelectOptionFeeder<T> {
 	scrollTo(index: number);
 
 }
+
+
+/**
+ * The component that provides available options for lu-select
+ */
+
+export abstract class ASelectOptionFeeder<T> implements ISelectOptionFeeder<T> {
+
+
+	protected _callbackKeyEvent: (event: KeyboardEvent) => void;
+	protected _callbackOptions: (options: LuSelectOption<T>[]) => void;
+	/**
+	 * See ISelectOptionFeeder
+	*/
+	open(): void {}
+	/**
+	 * See ISelectOptionFeeder
+	 */
+	hasFocus(): boolean {
+		return false;
+	}
+
+	/**
+	 * See ISelectOptionFeeder
+	 */
+	registerKeyevent(callback: (event: KeyboardEvent) => void): void {
+		this._callbackKeyEvent = callback;
+	}
+
+	/**
+	 * See ISelectOptionFeeder
+	 */
+	registerChangeOptions(callback: (options: LuSelectOption<T>[]) => void): void {
+		this._callbackOptions = callback;
+	}
+
+	/**
+	 * See ISelectOptionFeeder
+	 */
+	scrollTo(index: number) {
+	}
+
+}
