+++ packages/ng/src/app/select/option/feeder/select-option-feeder.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { LuOptionFeederDirective } from './select-option-feeder.directive';
-import { AbstractSelectOptionFeederComponent } from './select-option-feeder.component';
 
 @NgModule({
 	imports: [
@@ -9,11 +8,9 @@ import { AbstractSelectOptionFeederComponent } from './select-option-feeder.comp
 	],
 	declarations: [
 		LuOptionFeederDirective,
-		AbstractSelectOptionFeederComponent,
 	],
 	exports: [
 		LuOptionFeederDirective,
-		AbstractSelectOptionFeederComponent,
 	]
 })
 export class LuSelectOptionFeederModule { }
